requirements.txt
torch==1.7.0
hydra-core==1.0.3
loguru==0.5.3
numpy==1.18.5
omegaconf==2.0.2
tqdm==4.46.0
Cython==0.29.21
termcolor==1.1.0
scipy==1.5.2
